Rename controllers to handlers.
